添加角色相关的接口

+++ soho-admin/soho-admin-biz/src/main/java/work/soho/admin/mapper/AdminRoleMapper.java
@@ -13,5 +13,5 @@ public interface AdminRoleMapper extends BaseMapper<AdminRole> {
      *
      * @return
      */
-    List<AdminRole> getAdminRoleListByUserId(Integer userId);
+    List<AdminRole> getAdminRoleListByUserId(Long userId);
 }
